# packages/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Broker MVP API
  version: 0.1.0
  description:
    Minimal OpenAPI for the broker MVP. Data-plane endpoints require mutual TLS and a short-lived session token.
servers:
  - url: https://broker.example
security: []
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    Error:
      type: object
      required: [error, message, correlation_id]
      properties:
        error:
          type: string
        message:
          type: string
        correlation_id:
          type: string
      additionalProperties: false
    AdminRole:
      type: string
      enum: [owner, admin, auditor, operator]
...
    AdminUserStatus:
      type: string
      enum: [active, pending, disabled]
    AdminAccessRequestStatus:
      type: string
      enum: [pending, approved, denied, canceled]
    AdminUser:
      type: object
      required: [identity_id, issuer, subject, email, status, roles, tenant_ids, created_at, updated_at]
      properties:
        identity_id:
          type: string
        issuer:
          type: string
          format: uri
        subject:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        status:
          $ref: '#/components/schemas/AdminUserStatus'
        roles:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AdminRole'
        tenant_ids:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    AdminUserListResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        next_cursor:
          type: string
      additionalProperties: false
    AdminUserUpdateRequest:
      type: object
      minProperties: 1
      properties:
        status:
          $ref: '#/components/schemas/AdminUserStatus'
        roles:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AdminRole'
        tenant_ids:
          type: array
          items:
            type: string
      additionalProperties: false
    AdminAccessRequest:
      type: object
      required:
        [request_id, issuer, subject, email, requested_roles, requested_tenant_ids, status, created_at, updated_at]
      properties:
        request_id:
          type: string
        issuer:
          type: string
          format: uri
        subject:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        requested_roles:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AdminRole'
        requested_tenant_ids:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/AdminAccessRequestStatus'
        reason:
          type: string
        decided_by:
          type: string
        decided_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    AdminAccessRequestListResponse:
      type: object
      required: [requests]
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/AdminAccessRequest'
        next_cursor:
          type: string
      additionalProperties: false
    AdminAccessRequestCreateRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
      additionalProperties: false
    AdminAccessRequestApproveRequest:
      type: object
      properties:
        roles:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AdminRole'
        tenant_ids:
          type: array
          items:
            type: string
        reason:
          type: string
          minLength: 1
      additionalProperties: false
    AdminAccessRequestDenyRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
      additionalProperties: false
...
paths:
...
  /v1/admin/users:
    get:
      summary: List admin users for management
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AdminUserStatus'
        - name: tenant_id
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AdminRole'
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Admin users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/admin/users/{identityId}:
    patch:
      summary: Update admin user status, roles, and tenant scope
      security:
        - bearerAuth: []
      parameters:
        - name: identityId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdateRequest'
      responses:
        '200':
          description: Updated admin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          description: Invalid admin user update request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an admin user (disables access)
      security:
        - bearerAuth: []
      parameters:
        - name: identityId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Admin user deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Admin user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/admin/access-requests:
    post:
      summary: Submit an admin access request for owner review
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAccessRequestCreateRequest'
      responses:
        '201':
          description: Admin access request submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccessRequest'
        '400':
          description: Invalid access request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access request blocked by policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Access request cannot be created in current identity state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List admin access requests for review
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AdminAccessRequestStatus'
        - name: tenant_id
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AdminRole'
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Admin access requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccessRequestListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/admin/access-requests/{requestId}:
    get:
      summary: Get an admin access request by id
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Admin access request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccessRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Access request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/admin/access-requests/{requestId}/approve:
    post:
      summary: Approve admin access request and optionally override roles/tenant scopes
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAccessRequestApproveRequest'
      responses:
        '200':
          description: Updated admin access request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccessRequest'
        '400':
          description: Invalid approval request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Access request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/admin/access-requests/{requestId}/deny:
    post:
      summary: Deny admin access request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminAccessRequestDenyRequest'
      responses:
        '200':
          description: Updated admin access request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAccessRequest'
        '400':
          description: Invalid denial request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Access request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...